+++ app/controllers/Files.scala
@@ -94,7 +94,7 @@ class Files @Inject() (
             if (previewer.file)
             if (!file.showPreviews.equals("None")) && (previewer.contentType.contains(file.contentType))
           ) yield {
-            val tabTitle = "Preview"
+            val tabTitle = previewer.id
             if (file.licenseData.isDownloadAllowed(user) ||
               Permission.checkPermission(user, Permission.DownloadFiles, ResourceRef(ResourceRef.file, file.id))) {
               (file.id.toString, previewer.id, previewer.path, previewer.main, routes.Files.file(file.id) + "/blob",
@@ -202,6 +202,8 @@ class Files @Inject() (
         val isRDFExportEnabled = current.plugin[RDFExportService].isDefined
 
         val extractionsByFile = extractions.findById(new ResourceRef('file, id))
+        val extractionGroups = extractions.groupByType(extractionsByFile)
+
 
         var folderHierarchy = new ListBuffer[Folder]()
         if (foldersContainingFile.length > 0) {
@@ -236,7 +238,7 @@ class Files @Inject() (
             plugin.getOutputFormats(contentTypeEnding).map(outputFormats =>
               Ok(views.html.file(file, id.stringify, commentsByFile, previewsWithPreviewer, sectionsWithPreviews,
                 extractorsActive, decodedDatasetsContaining.toList, foldersContainingFile,
-                mds, isRDFExportEnabled, extractionsByFile, outputFormats, space, access, folderHierarchy.reverse.toList, decodedSpacesContaining.toList, allDecodedDatasets.toList, view_count, view_date)))
+                mds, isRDFExportEnabled, extractionGroups, outputFormats, space, access, folderHierarchy.reverse.toList, decodedSpacesContaining.toList, allDecodedDatasets.toList, view_count, view_date)))
           }
           case None =>
             Logger.debug("Polyglot plugin not found")
@@ -247,7 +249,7 @@ class Files @Inject() (
             //passing None as the last parameter (list of output formats)
             Future(Ok(views.html.file(file, id.stringify, commentsByFile, previewsWithPreviewer, sectionsWithPreviews,
               extractorsActive, decodedDatasetsContaining.toList, foldersContainingFile,
-              mds, isRDFExportEnabled, extractionsByFile, None, space, access, folderHierarchy.reverse.toList, decodedSpacesContaining.toList, allDecodedDatasets.toList, view_count, view_date)))
+              mds, isRDFExportEnabled, extractionGroups, None, space, access, folderHierarchy.reverse.toList, decodedSpacesContaining.toList, allDecodedDatasets.toList, view_count, view_date)))
         }
       }